Fixed rather stupid bug
The "wrongs" in the training loop were computed incorrectly. I wondered why training does not converge - that's why. After I inserted test code for train data I found that bug and fixed it.

I should not write code in the half a hour when son sleeping and I am not yet.